pyproject.toml
[tool.poetry]
name = "repl-nix-yalab-rrt"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "repl_nix_yalab_rrt"}]

[tool.poetry.dependencies]
python = "^3.10"
flask-cors = "^4.0.1"
pandas = "^2.2.2"
flask = "^3.0.3"
openpyxl = "3.1.5"
oauth2client = "^4.1.3"
gspread = "^6.1.2"
google-api-python-client = "^2.137.0"
google-auth = "^2.32.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
